Refactorizar codigo
