+++ src/test/java/Presidencia/Test/RunnersFeature.java
@@ -5,7 +5,7 @@ import cucumber.api.CucumberOptions;
 import net.serenitybdd.cucumber.CucumberWithSerenity;
 
 @RunWith(CucumberWithSerenity.class)
-@CucumberOptions(features = { "src/test/resources/features" }, glue = { "Presidencia.Definitions" }, tags = {"@Regresion"})
+@CucumberOptions(features = { "src/test/resources/features" }, glue = { "Presidencia.Definitions" })
 
 public class RunnersFeature {
