+++ src/test/resources/features/sprint 1/1. PeriodoCumplimiento.feature
@@ -1,11 +1,14 @@
 #30/04/2020
 #Jhonatan Medina Blanco
-@Regresion
+@Regresion 
 Feature: Registro del Período para la Gestión del Cumplimiento. Hu(007) (USER STORY 131817) 
-Scenario Outline: Registro del Período para la Gestión del Cumplimiento (132981) 
+Background: 
 	Given que se ingreso a la url 
+Scenario Outline: Registro del Período para la Gestión del Cumplimiento (132981) 
+
 	When Se consulte la fecha inicio <FechaInicio> y fecha fin <FechaFin> 
 	And se debe anexar documentos y ver archivos cargados 
+	Then al hacer clic en guardar se debe almacenar la informacion
 	And Se cierra el navegador 
 	
 	Examples:
